refactored graphql folder and started on CRUD operations for Job
