+++ src/graphql/index.ts
@@ -1,1 +1,4 @@
+export * from './Job';
+export * from './Mutation';
+export * from './Query';
 export * from './User';
